requirements.txt
# ============================================
# Confluence RAG 시스템 - 프로덕션 의존성
# ============================================

# 웹 크롤링 (Confluence 페이지 자동 수집)
playwright==1.40.0

# HTML 파싱 (크롤링한 페이지 구조 분석)
beautifulsoup4==4.12.2

# HTML -> Markdown 변환 (문서 정제)
markdownify==0.11.6

# RAG 프레임워크 (검색 증강 생성 파이프라인)
langchain==0.1.0

# 벡터 데이터베이스 (임베딩 저장 및 유사도 검색)
chromadb==0.4.22

# 임베딩 모델 (텍스트를 벡터로 변환)
sentence-transformers==2.2.2

# 웹 UI (사용자 질의응답 인터페이스)
gradio==4.16.0

# 환경변수 관리 (.env 파일 로드)
python-dotenv==1.0.0

# CLI 인터페이스 (명령줄 도구 구성)
click==8.1.7

# 진행바 표시 (크롤링/임베딩 진행 상황)
tqdm==4.66.1

# 터미널 출력 포맷팅 (로그, 테이블 등 예쁜 출력)
rich==13.7.0
